move `ensure_no_collision` to `bevy_macro_utils`
