+++ crates/bevy_ecs/macros/src/fetch.rs
@@ -1,3 +1,4 @@
+use bevy_macro_utils::ensure_no_collision;
 use proc_macro::TokenStream;
 use proc_macro2::{Ident, Span};
 use quote::{quote, ToTokens};
@@ -107,10 +108,10 @@ pub fn derive_world_query_impl(input: TokenStream) -> TokenStream {
     };
 
     let fetch_struct_name = Ident::new(&format!("{struct_name}Fetch"), Span::call_site());
-    let fetch_struct_name = crate::ensure_no_collision(tokens.clone(), fetch_struct_name);
+    let fetch_struct_name = ensure_no_collision(fetch_struct_name, tokens.clone());
     let read_only_fetch_struct_name = if fetch_struct_attributes.is_mutable {
         let new_ident = Ident::new(&format!("{struct_name}ReadOnlyFetch"), Span::call_site());
-        crate::ensure_no_collision(tokens.clone(), new_ident)
+        ensure_no_collision(new_ident, tokens.clone())
     } else {
         fetch_struct_name.clone()
     };
@@ -118,7 +119,7 @@ pub fn derive_world_query_impl(input: TokenStream) -> TokenStream {
     // Generate a name for the state struct that doesn't conflict
     // with the struct definition.
     let state_struct_name = Ident::new(&format!("{struct_name}State"), Span::call_site());
-    let state_struct_name = crate::ensure_no_collision(tokens, state_struct_name);
+    let state_struct_name = ensure_no_collision(state_struct_name, tokens);
 
     let fields = match &ast.data {
         Data::Struct(DataStruct {
